# frontend/openapi/food-helper-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FoodHelper.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5099/"
    }
  ],
  "paths": {
...
    "/api/ingredient-words": {
      "get": {
        "tags": [
          "IngredientWords"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientWord"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientWord"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientWord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "IngredientWords"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIngredientWordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIngredientWordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddIngredientWordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientWord"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientWord"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientWord"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredient-words/{id}": {
      "delete": {
        "tags": [
          "IngredientWords"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddIngredientWordRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "IngredientWord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}